+++ Packages/com.projectdawn.navigation/ProjectDawn.Navigation.Astar/AgentAstarPath.cs
@@ -1,6 +1,7 @@
 #if ENABLE_ASTAR_PATHFINDING_PROJECT
 using Unity.Entities;
 using Pathfinding.ECS;
+using UnityEngine;
 
 namespace ProjectDawn.Navigation.Astar
 {
@@ -13,7 +14,14 @@ namespace ProjectDawn.Navigation.Astar
         /// <summary>
         /// Policy for how often to recalculate an agent's path.
         /// </summary>
+        [Tooltip("Policy for how often to recalculate an agent's path.")]
         public AutoRepathPolicy AutoRepath;
+
+        /// <summary>
+        /// Controls how agent will be grounded to the surface.
+        /// </summary>
+        [Tooltip("Controls how agent will be grounded to the surface.")]
+        public Grounded Grounded;
 
         /// <summary>
         /// Returns default configuration.
@@ -21,6 +29,7 @@ namespace ProjectDawn.Navigation.Astar
         public static AgentAstarPath Default => new()
         {
             AutoRepath = AutoRepathPolicy.Default,
+            Grounded = Grounded.XYZ,
         };
     }
 }