+++ Packages/com.projectdawn.navigation/ProjectDawn.Navigation.Hybrid/AgentReciprocalAvoidAuthoring.cs
@@ -1,5 +1,6 @@
 using Unity.Entities;
 using UnityEngine;
+using ProjectDawn.Entities;
 
 namespace ProjectDawn.Navigation.Hybrid
 {
@@ -15,6 +16,13 @@ namespace ProjectDawn.Navigation.Hybrid
         protected NavigationLayers m_Layers = NavigationLayers.Everything;
 
         Entity m_Entity;
+
+        /// <summary>
+        /// <see cref="AgentReciprocalAvoid"/> component of this <see cref="AgentAuthoring"/> Entity.
+        /// Accessing this property is potentially heavy operation as it will require wait for agent jobs to finish.
+        /// </summary>
+        public ref AgentReciprocalAvoid Sonar =>
+            ref World.DefaultGameObjectInjectionWorld.EntityManager.GetComponentDataRW<AgentReciprocalAvoid>(m_Entity).ValueRW;
 
         /// <summary>
         /// Returns default component of <see cref="AgentReciprocalAvoid"/>.