+++ Packages/com.projectdawn.navigation/ProjectDawn.Navigation/Grounded.cs
@@ -1,0 +1,9 @@
+namespace ProjectDawn.Navigation
+{
+    public enum Grounded : byte
+    {
+        None,
+        XYZ,
+        XZ,
+    }
+}